+++ plugins/multi_domain/files/apps/sites/models.py
@@ -3,7 +3,7 @@ import re
 from django.conf import settings
 from django.contrib.sites.models import Site
 from django.db import models
-from django.utils.html import mark_safe
+from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 
 from apps.utils.models import BaseModel